requirements.txt
# Minimal requirements for the GUI demo + graph view and numpy-based pieces
streamlit>=1.36
numpy>=1.26
networkx>=3.2
pyvis>=0.3.2
pandas
pydantic
pytest
python-docx
openai>=1.0.0
python-dotenv
pdfplumber
python-pptx
